Stop leaking padding bytes into world snapshots
Empty marker components (Hidden/Oneway/NoBounce) memcpy'd their single
uninitialized object byte into snapshots, and CampaignSystemTag's mixed
int/double layout carried uninitialized padding — checkpoint hashes
diverged between identical runs and replays falsely flagged divergence.
Markers now encode a fixed byte (decoders ignore the payload, so old
saves still load), padded structs are reordered to be padding-free, and
encode_pod/encode_tag statically reject empty types.

Host replays also get --replay-exit (0 on replay_verified, 1 on
divergence), the journal tick now advances every frame instead of only
under --frames, journaled runs step deterministically once per frame,
and replay without --fixed-hz prints a determinism warning.

+++ engine/include/stellar/engine/scene_components.hpp
@@ -165,8 +165,11 @@ struct DoubleSided {
 // resolved world entity so F5/F9 snapshots restore the camera too (the
 // document seeds it only on scene load).
 struct Camera3DState {
+  // All-double members keep the layout padding-free — the snapshot codec
+  // memcpy's the object representation and padding bytes would leak
+  // uninitialized memory into byte-compared snapshots.
   double x{}, y{}, z{3.0};
-  float yaw_deg{}, pitch_deg{}, fov_deg{60.f};
+  double yaw_deg{}, pitch_deg{}, fov_deg{60.0};
 };
 // 3D positional attachment — same contract as Parent, with a z offset.
 struct Parent3D {
